Move advanced_snippet_validator
